rename: Reorganize project folder structure
- 기능 수정 없음
