+++ app/src/main/java/campus/tech/kakao/map/repository/PreferenceRepository.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.repository
 
 import android.content.Context
 import android.content.SharedPreferences
